Program that converts video and predicts labels
Python class Emotion_ML() has two functions:  
- vid_to_imgs(self, file_name) 
- predict_emotions(self, img_dir)

vid_to_imgs(self file_name) takes in the path of the video file and uses the video to convert it into frames for the model. Each image is named after its time stamp in the video and saved to an 'imgs' folder.

predict_emotions(self, img_dir) takes in the imgs folder directory made in the previous function to add them to an np array, reshape it to (-1, 48, 48, 3) for the model, and then preprocesses the array for MobileNetV2. This function also imports the model, which the np array is applied to in order to get the prediction labels.